requirements.txt
absl-py==0.9.0
annoy==1.16.3
astunparse==1.6.3
boto3==1.7.4
botocore==1.10.84
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
docutils==0.16
Flask==1.1.2
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.0
future==0.18.2
gast==0.3.3
gevent==1.4.0
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.30.0
gunicorn==19.9.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
Keras==2.3.1
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
nltk==3.5
numpy==1.18.0
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.0.3
Pillow==6.2.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2020.1
psycopg2==2.9.4
PyYAML==5.3.1
rank-bm25==0.2.1
redis==3.5.3
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.1.13
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.4.1
sentence-transformers==2.0.0
sentencepiece==0.1.91
six==1.15.0
SQLAlchemy==1.3.18
sqlalchemy-mixins==1.2.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-hub==0.8.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.9.0
tqdm==4.48.0
transformers==4.9.2
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0